Split up flow-graph-tests.ts into smaller files.
