+++ src/dataflow/analysis/testing/flow-graph-testing.ts
@@ -1,0 +1,23 @@
+/**
+ * @license
+ * Copyright 2019 Google LLC.
+ * This code may only be used under the BSD style license found at
+ * http://polymer.github.io/LICENSE.txt
+ * Code distributed by Google as part of this project is also
+ * subject to an additional IP rights grant found at
+ * http://polymer.github.io/PATENTS.txt
+ */
+
+import {assert} from '../../../platform/chai-web.js';
+import {Manifest} from '../../../runtime/manifest.js';
+import {FlowGraph} from '../flow-graph.js';
+
+/** Constructs a FlowGraph from the recipe in the given manifest. */
+export async function buildFlowGraph(manifestContent: string): Promise<FlowGraph> {
+  const manifest = await Manifest.parse(manifestContent);
+  assert.lengthOf(manifest.recipes, 1);
+  const recipe = manifest.recipes[0];
+  assert(recipe.normalize(), 'Failed to normalize recipe.');
+  assert(recipe.isResolved(), 'Recipe is not resolved.');
+  return new FlowGraph(recipe);
+}
